Culmination of work
I still don't know how to properly make branches seperated from each
other, and I failed a lot in the beggining so I ended up deleting
everything and starting from scratch even though we were told not to
but I didn't ask for help until the day before this was due.

Lesson learned.

the project is not complete but this is what I have so far.

night